Update multiple project configurations and dependencies
The changes in this commit include an update to the `.gitignore` file to ignore Jetbrains files, a tweak in the formatting and types used in various TypeScript files, and an augmentation of the ESLint configuration in `.eslintrc.json` to add further environment settings, extended configurations, parser options, and added plugins. The `package-lock.json` and `package.json` files have been updated to install new dependencies and update existing ones.