Add Jest testing framework and update tsconfig.json and eslint configurations
This commit introduces Jest as the testing framework into the project, installs new testing-related dependencies and adjusts tsconfig and ESLint configurations to accommodate these changes. Additionally, a new test for a container component has been added. The configurations updates improve the flexibility of the project to accommodate better testing practices.